Added new preprocessor directive checks using global scripting define symbols
This should make it more reliable to check for dynamic bones and the sdk
