+++ Resources/Presets/Blendshapes/test.asset
@@ -1,0 +1,251 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 65df42acc29efd546902636e201a2390, type: 3}
+  m_Name: test
+  m_EditorClassIdentifier: 
+  name: test
+  renderers:
+  - rendererPath: Body
+    blendshapes:
+    - name: vrc.blink_left
+      weight: 100
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.blink_right
+      weight: 100
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.lowerlid_left
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.lowerlid_right
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_aa
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_ch
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_dd
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_e
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_ff
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_ih
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_kk
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_nn
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_oh
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_ou
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_pp
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_rr
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_sil
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_ss
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: vrc.v_th
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Blink
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Wink
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Wink Right
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Blink Happy
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Wink 2
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Wink 2 Right
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: '> <'
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Surprised
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Stare
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Ah
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Ch
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: U
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: E
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Oh
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: "\u25B3"
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: What
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: "\u03C9"
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Smile
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: "\u3002"
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Wa
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: E What
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Mumu
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: "\u25A1"
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: ~
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: The Body It is
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Pupil
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Circle White Eye
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Round and round Eye
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Highlight Eliminate
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Ch Take
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Teeth Eliminate
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Cheerful
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Sadness
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Anger
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Eyebrow Lower
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Cheek Blush
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Cold Sweat
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: Blue Give up
+      weight: 0
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    - name: HatOn
+      weight: 100
+      otherNames: []
+      otherNamesExpandedInUI: 0
+    expandedInUI: 0
