+++ Resources/Presets/Poses/weird 2.asset
@@ -1,0 +1,114 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8d5ce1fd657d8984c8ced97b71ea3c84, type: 3}
+  m_Name: weird 2
+  m_EditorClassIdentifier: 
+  name: weird 2
+  muscles:
+  - -0.0000007683962
+  - -0
+  - -0
+  - -0.0000005122642
+  - -7.633331e-15
+  - -0.000000042688683
+  - 0
+  - 0
+  - 0
+  - 0.58999956
+  - -0.59000015
+  - -0.23000008
+  - 0.0000013660377
+  - -0.0000000853774
+  - 0.00000008537736
+  - -0
+  - -0
+  - 0.019999983
+  - 0
+  - 0
+  - 0
+  - 0.599809
+  - 0.0061013685
+  - -0.0032078803
+  - 0.99964255
+  - 0.0009155491
+  - 0.8799939
+  - 0.0047992817
+  - 0.00000054641504
+  - -0.880001
+  - 0.0062515023
+  - -0.0031205537
+  - 0.99964124
+  - 0.0017112991
+  - -0.00013920703
+  - 0.005896454
+  - -3.053331e-15
+  - 2.0000007
+  - 0.72999996
+  - -0.9104499
+  - 0.32365933
+  - -0.91401887
+  - 0.8148458
+  - 1.4148241
+  - 0.3261323
+  - 0.05845393
+  - 0.9
+  - -0.9000006
+  - 0.5089729
+  - 0.27209902
+  - -0.08102567
+  - 0.60245854
+  - 0.097906806
+  - 0.003914115
+  - 0.01717775
+  - -1.0226737
+  - 0.14928113
+  - 0.14642537
+  - 0.63732153
+  - 0.6672432
+  - -0.4636938
+  - 0.81089985
+  - 0.8109
+  - 0.66711706
+  - -0.64282054
+  - 0.81146735
+  - 0.81146806
+  - 0.66509134
+  - -0.69068503
+  - 0.8110472
+  - 0.8110474
+  - 0.6611733
+  - -0.49860105
+  - 0.81589633
+  - 0.81008077
+  - -1.0906131
+  - 0.119029626
+  - 0.31701288
+  - 0.6388712
+  - 0.6667738
+  - -0.46729556
+  - 0.81111825
+  - 0.8111174
+  - 0.6668068
+  - -0.65158087
+  - 0.81157255
+  - 0.811572
+  - 0.6654669
+  - -0.68306357
+  - 0.81123924
+  - 0.81123924
+  - 0.6617345
+  - -0.4959907
+  - 0.8162026
+  - 0.81042
+  presetMode: 0
+  transformPaths: []
+  transformRotations: []
