+++ Scripts/Editor/Presets/PumkinsCameraPreset.cs
@@ -1,5 +1,6 @@
 ﻿using Pumkin.AvatarTools;
 using Pumkin.DataStructures;
+using Pumkin.Dependencies;
 using Pumkin.HelperFunctions;
 using System;
 using UnityEditor;
