+++ Scripts/Editor/Presets/PumkinsPosePreset.cs
@@ -1,5 +1,6 @@
 ﻿using Pumkin.AvatarTools;
 using Pumkin.DataStructures;
+using Pumkin.Dependencies;
 using Pumkin.HelperFunctions;
 using System;
 using System.Collections;
